Add level ups messages
Thisincludes a big refactor on how experience is edited for hunters. Now, we use a specific `edit_experience_with_levelups` function so that we can send level up messages properly without delaying the bot on save and messing with shady internals to store the initial experience.

The downside is that you shouldn't use the experience value directly anymore if you want to edit an hunter experience.