Update the landmines game to add a workaround for https://github.com/tortoise/tortoise-orm/issues/822
member, a OneToOneField, is no longer the LandminesUserData primary key because tortoise fails to generate correct SQL queries when running with a OneToOneField as a primary_key.

There will be an available optimisation if/when a fix for https://github.com/tortoise/tortoise-orm/issues/822 gets released: remove the implicit id column and use member as the table primary key again.